fix(monitors): Match error detail breadcrumbs to edit
Use the same project-aware top bar breadcrumb on the error monitor detail and edit pages when page-frame is enabled.

Keep the legacy header path unchanged so this stays scoped to the new top bar layout.
